añadidos loggers

+++ src/main/java/com/nova/everisdarmytasksms/controller/TaskController.java
@@ -2,6 +2,8 @@ package com.nova.everisdarmytasksms.controller;
 
 import java.util.List;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -27,6 +29,8 @@ public class TaskController {
 	@Autowired
 	TaskService taskService;
 	
+	private static final Logger logger =  LoggerFactory.getLogger(TaskController.class);
+	
 	//Return all the tasks stored in DB
 	@GetMapping("/tasks")
 	public List<Task> getTasks() {
@@ -39,9 +43,11 @@ public class TaskController {
 			@RequestParam("description") String description) {
 		Task task = taskService.createTask(status, description);
 		if (task != null) {
+			logger.info("Se puede crear la tarea correctamente");
 			taskRepository.save(task);
 			return new ResponseEntity<>("Tarea creada correctamente", HttpStatus.CREATED);
 		} else {
+			logger.info("La tarea no ha podido ser creada");
 			return new ResponseEntity<>("La tarea debe tener una descripción "
 					+ "con una longitud máxima de 255 caracteres", 
 					HttpStatus.UNPROCESSABLE_ENTITY);
@@ -60,6 +66,7 @@ public class TaskController {
 			@RequestBody Task task) {
 		Task existingTask = taskService.getTaskById(id);
 		if (existingTask != null) {
+			logger.info("La tarea con el id " + id + " ha sido actualizada");
 			existingTask.setDescription(task.getDescription());
 			existingTask.setStatus(task.getStatus());
 			taskRepository.save(existingTask);
@@ -67,6 +74,7 @@ public class TaskController {
 			return new ResponseEntity<Task>(existingTask, 
 					HttpStatus.OK);
 		} else {
+			logger.info("La tarea con el id " + id + " no existe");
 			return new ResponseEntity<Task>(HttpStatus.BAD_REQUEST);
 		}
 	}
@@ -77,12 +85,14 @@ public class TaskController {
 			@RequestParam("description") String description) {
 		Task existingTask = taskService.getTaskById(id);
 		if (existingTask != null) {
+			logger.info("La descripcion de la tarea con el id " + id + " ha sido actualizada");
 			existingTask.setDescription(description);
 			taskRepository.save(existingTask);
 			
 			return new ResponseEntity<Task>(existingTask, 
 					HttpStatus.OK);
 		} else {
+			logger.info("La tarea con el id " + id + " no existe");
 			return new ResponseEntity<Task>(HttpStatus.BAD_REQUEST);
 		}
 	}
@@ -93,12 +103,14 @@ public class TaskController {
 			@RequestParam("status") String status) {
 		Task existingTask = taskService.getTaskById(id);
 		if (existingTask != null) {
+			logger.info("El status de la tarea con el id " + id + " ha sido actualizada");
 			existingTask.setStatus(status);
 			taskRepository.save(existingTask);
 			
 			return new ResponseEntity<Task>(existingTask, 
 					HttpStatus.OK);
 		} else {
+			logger.info("La tarea con el id " + id + " no existe");
 			return new ResponseEntity<Task>(HttpStatus.BAD_REQUEST);
 		}
 	}
@@ -106,9 +118,16 @@ public class TaskController {
 	//Delete the task with the given id
 	@DeleteMapping("/tasks/{id}")
 	public ResponseEntity<String> deleteTask(@PathVariable("id") Integer id) {
-		taskRepository.deleteById(id);
+		if (taskRepository.existsById(id)) {
+			taskRepository.deleteById(id);
+			logger.info("La tarea con el id " + id + " ha sido eliminada correctamente");
+			return new ResponseEntity<String>("Tarea eliminada correctamente", 
+					HttpStatus.OK);
+		} else {
+			logger.info("La tarea con el id " + id + " no existe");
+			return new ResponseEntity<String>("La tarea con el id indicado no existe", HttpStatus.BAD_REQUEST);
+		}
 		
-		return new ResponseEntity<String>("Tarea eliminada correctamente", 
-				HttpStatus.OK);
+		
 	}
 }
